Move slurm and memory tools under src/tools/
First half of collapsing the tools namespace. Living inside src/ means
tools inherit the coverage, lint, and test conventions structurally
rather than through per-directory configuration.

Both tools are self-contained and stay that way: parse.py models
Snakemake benchmark TSVs and SLURM .err files, and the memory estimator
models pyoxigraph RSS. Neither touches Babel data, so neither has a
natural home in src/ proper.

tools/memory/README.md becomes docs/tools/Memory.md, so every tool's
documentation is reachable from the one index.

This also retires the tests/README.md rule forbidding
tests/tools/__init__.py. That rule existed only because a test package
named `tools` shadowed the real top-level `tools` package. With no
top-level `tools`, tests/tools/ can be a normal package like its
siblings.

tools/__init__.py is kept for now; tools/clique_diff still needs it and
is moved in a later commit.
