+++ pyproject.toml
@@ -32,8 +32,8 @@ dependencies = [
 ]
 
 [project.scripts]
-babel-slurm-errors = "tools.slurm.errors:main"
-babel-slurm-resources = "tools.slurm.resources:main"
+babel-slurm-errors = "src.tools.slurm.errors:main"
+babel-slurm-resources = "src.tools.slurm.resources:main"
 babel-clique-diff = "tools.clique_diff.diff:main"
 source-impact-report = "src.cli.source_impact_report:main"
