+++ src/tools/slurm/resources.py
